moved jdk util into main package
