+++ benchmarks/src/main/java/us/hebi/quickbuf/benchmarks/jdk/JdkMethodsBenchmark.java
@@ -25,7 +25,7 @@ import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
 import org.openjdk.jmh.runner.options.VerboseMode;
-import us.hebi.quickbuf.jdk.JdkMath;
+import us.hebi.quickbuf.JdkMath;
 
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
