+++ runtime/src/java9/compile.bat
@@ -1,3 +1,3 @@
 @echo off
 SET mypath=%~dp0
-"%JAVA11%/bin/javac" -source 6 -target 6 %mypath:~0,-1%/JdkMethods.java
+"%JAVA11%/bin/javac" -source 6 -target 6 %mypath:~0,-1%/*.java
